+++ src/domains/Home/components/SearchBar/SearchBar.style.ts
